refactor: centralize Sentry imports through @vendor/observability and misc fixes
- Add @vendor/observability/sentry re-export of @sentry/core
- Migrate backfill, gateway, relay from direct @sentry/core to vendor abstraction
- Remove @sentry/core direct dependency from service apps
- Extract timingSafeStringEqual into backfill crypto utility
- Fix agent empty messages nullish check in ai-sdk
- Handle BetterStack log shipping errors with .catch()
