+++ apps/gateway/src/middleware/sentry.ts
@@ -1,4 +1,4 @@
-import * as Sentry from "@sentry/core";
+import { captureException, captureMessage, withScope } from "@vendor/observability/sentry";
 import { createMiddleware } from "hono/factory";
 import type { RequestIdVariables } from "./request-id.js";
 
@@ -16,7 +16,7 @@ export const sentry = createMiddleware<{
   try {
     await next();
   } catch (err) {
-    Sentry.withScope((scope) => {
+    withScope((scope) => {
       scope.setTag("service", "gateway");
       scope.setTag("http.method", c.req.method);
       scope.setTag("http.path", c.req.path);
@@ -30,13 +30,13 @@ export const sentry = createMiddleware<{
         requestId,
         correlationId,
       });
-      Sentry.captureException(err);
+      captureException(err);
     });
     throw err;
   }
 
   if (c.res.status >= 500) {
-    Sentry.withScope((scope) => {
+    withScope((scope) => {
       scope.setTag("service", "gateway");
       scope.setTag("http.method", c.req.method);
       scope.setTag("http.path", c.req.path);
@@ -51,7 +51,7 @@ export const sentry = createMiddleware<{
         requestId,
         correlationId,
       });
-      Sentry.captureMessage(
+      captureMessage(
         `HTTP ${c.res.status}: ${c.req.method} ${c.req.path}`,
         "error",
       );
