+++ apps/relay/package.json
@@ -24,7 +24,6 @@
     "@repo/console-validation": "workspace:*",
     "@repo/gateway-types": "workspace:*",
     "@repo/lib": "workspace:*",
-    "@sentry/core": "catalog:",
     "@t3-oss/env-core": "catalog:",
     "@vendor/db": "workspace:*",
     "@vendor/observability": "workspace:*",
